backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.9.2
numpy==2.1.1
pillow==10.4.0
